variadic_from: refactoring
